requirements.txt
amqp==2.5.2
appdirs==1.4.3
astroid==2.4.0
attrs==19.3.0
billiard==3.6.3.0
black==19.10b0
cachetools==4.0.0
celery==4.4.2
certifi==2019.11.28
chardet==3.0.4
Click==7.0
clickclick==1.2.2
codacy-coverage==1.3.11
connexion==2.7.0
coverage==5.1
distlib==0.3.0
dnspython==1.16.0
docker==4.2.0
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Executor==0.9.3
flask-marshmallow==0.11.0
Flask-SQLAlchemy==2.3.2
flatten-json==0.1.6
gitdb==4.0.5
GitPython==3.1.7
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
gunicorn==20.0.4
httplib2==0.18.0
idna==2.9
importlib-metadata==1.5.0
importlib-resources==1.5.0
inflection==0.3.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
kombu==4.6.8
lazy-object-proxy==1.4.3
MarkupSafe==1.1.0
marshmallow==3.5.0
marshmallow-oneofschema==2.0.1
marshmallow-sqlalchemy==0.22.3
mccabe==0.6.1
more-itertools==8.2.0
mysql-connector-python==8.0.19
mysqlclient==1.4.6
openapi-spec-validator==0.2.8
packaging==20.3
pathspec==0.8.0
pluggy==0.13.1
pre-commit==2.9.3
protobuf==3.6.1
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.5.0
PyMySQL==0.9.3
pyparsing==2.4.7
pyrsistent==0.15.7
pytest==5.4.1
pytz==2020.1
PyYAML==5.3
regex==2020.5.14
requests==2.23.0
rsa==4.0
six==1.12.0
smmap==3.0.4
SQLAlchemy==1.3.0
swagger-ui-bundle==0.0.6
toml==0.10.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
virtualenv==20.0.7
virtualenv-clone==0.5.3
wcwidth==0.1.9
websocket-client==0.57.0
Werkzeug==1.0.0
wrapt==1.12.1
zipp==3.1.0
